requirements.txt
websocket-client==0.35.0
six==1.10.0
requests==2.9.1
